Move factories around to make them accessible more widely
