+++ graph/test/test_srt.c
@@ -1,55 +1,7 @@
 #include <semaphore.h>
 #include "countdown.h"
 #include "srt.h"
-#include "pdls.c"
-#include "loader.h"
-
-typedef struct {
-    rage_ElementLoader * loader;
-    rage_ElementKind * kind;
-    rage_ConcreteElementType * cet;
-    rage_Element * elem;
-} rage_TestElem;
-
-typedef RAGE_OR_ERROR(rage_TestElem) rage_NewTestElem;
-
-void free_test_elem(rage_TestElem te) {
-    if (te.elem != NULL)
-        rage_element_free(te.elem);
-    if (te.cet != NULL)
-        rage_concrete_element_type_free(te.cet);
-    if (te.kind != NULL)
-        rage_element_loader_unload(te.kind);
-    if (te.loader != NULL)
-        rage_element_loader_free(te.loader);
-}
-
-rage_NewTestElem new_test_elem() {
-    rage_ElementLoader * el = rage_element_loader_new(getenv("RAGE_ELEMENTS_PATH"));
-    rage_TestElem rv = {.loader=el};
-    rage_ElementKindLoadResult ek_ = rage_element_loader_load("./libamp.so");
-    if (RAGE_FAILED(ek_)) {
-        free_test_elem(rv);
-        return RAGE_FAILURE_CAST(rage_NewTestElem, ek_);
-    }
-    rv.kind = RAGE_SUCCESS_VALUE(ek_);
-    rage_Atom ** tups = generate_tuples(rage_element_kind_parameters(rv.kind));
-    rage_NewConcreteElementType ncet = rage_element_type_specialise(
-        rv.kind, tups);
-    free_tuples(rage_element_kind_parameters(rv.kind), tups);
-    if (RAGE_FAILED(ncet)) {
-        free_test_elem(rv);
-        return RAGE_FAILURE_CAST(rage_NewTestElem, ncet);
-    }
-    rv.cet = RAGE_SUCCESS_VALUE(ncet);
-    rage_ElementNewResult elem_ = rage_element_new(rv.cet, 44100, 256);
-    if (RAGE_FAILED(elem_)) {
-        free_test_elem(rv);
-        return RAGE_FAILURE_CAST(rage_NewTestElem, elem_);
-    }
-    rv.elem = RAGE_SUCCESS_VALUE(elem_);
-    return RAGE_SUCCESS(rage_NewTestElem, rv);
-}
+#include "graph_test_factories.h"
 
 rage_Error test_srt() {
     rage_NewTestElem nte = new_test_elem();
